SourceTarball.bash: Fix case of in-module .git directory
Older Git versions keep the submodule .git directory inside the
submodule work tree instead of ".git/modules/<path>" in the outer
project.  On such versions the output of "git rev-parse --git-dir"
inside the submodule may be a relative path.  Convert it to a full path
before adding to GIT_ALTERNATE_OBJECT_DIRECTORIES so it is meaningful
outside the submodule directory.